% Please edit documentation in R/fetch_genome.R
\name{fetch.genome}
\alias{fetch.genome}
\title{Retrieve chromosomal length information from file}
\usage{
fetch.genome(genome = NULL)
}
\arguments{
\item{genome}{filename of the genome file (defaults to NULL); required. If unspecified, hg19 information is returned.}
}
\description{
Retrieves from a genome file (alternatively, a FASTA index file generated by samtools) a list of chromosomes and their corresponding lengths. If no input is specified, built-in hg19 information is returned as a data.frame.
}
\keyword{chromosome}
\keyword{genome}
\keyword{samtools}
