% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_which.r
\name{file.which}
\alias{file.which}
\title{Find path to an executable file}
\usage{
file.which(x)
}
\arguments{
\item{x}{filename of the executable file as a string; required.}
}
\description{
A utility functional analog of "which" that searches through $PATH for the first instance of a particular executable, or returns NULL if one cannot be found.
}
\examples{
file.which("R") # find your copy of R executable
file.which("Supercalifragilisticexpialidocious") # should return NULL unless you do have a file with this name
}
\keyword{which}
