% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natsort.R
\name{natsort}
\alias{natsort}
\title{Natural sorting}
\usage{
natsort(..., alg = NULL, coerce = NULL)
}
\arguments{
\item{alg}{path to an external sorter executable; defaults to the internal Perl version if unspecified.}
\item{coerce}{should values that look like numbers or NA's be converted likewise? defaults to FALSE.}
\item{...}{data to be passed to natsort() and sorted.}
}
\description{
Performs natural sorting via the alphanum algorithm on an input vector of characters. The algorithm employed here will give slightly different results than GNU sort (as invoked by sort -V) if the input begins with non-alphanumeric letters, e.g. literal ^, >, # characters.
}
\examples{
g <- c("1.1.11", NA, "02.3", "11.7v", "na")
sort(g)
# [1] "02.3"   "1.1.11" "11.7v"  "na"
natsort(g)
# [1] "1.1.11" "02.3"   "11.7v"  "na"     NA
natsort(g, alg = alg, coerce = TRUE)
# [1] "1.1.11" "02.3"   "11.7v"  NA       NA      
}
\keyword{perl}
\keyword{genome}