% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qw.R
\name{qw}
\alias{qw}
\title{Generate an array from a string of text}
\usage{
qw(x = "", as.is = TRUE, na.list = c("NA", "nan", "N/A", "n/a", "na"))
}
\arguments{
\item{x}{string input.}

\item{as.is}{keeps numeric entries as characters (defaults to TRUE); to coerce data into numeric values, set this to FALSE.}

\item{na.list}{character vector of strings that are equivalent to NA (case-sensitive).}
}
\description{
Similar to Perl's qw function, this array converts a whitespace-delimited string into a character or numeric array.
}
\examples{
qw("hello world! I can't believe this is not Perl!")
# [1] "hello" "world!" "I" "can't" "believe" "this" "is" "not" "Perl!"  
qw(1:10)
# [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"
qw(1:10, as.is = FALSE) # an integer vector from 1 to 10
# [1]  1  2  3  4  5  6  7  8  9 10
}
\keyword{qw}
