\name{rfPIPeak.setup}
\alias{rfPIPeak.setup}
\title{Post-installation configuration for rfPIPeak}
\usage{
rfPIPeak.setup(silent = TRUE)
}
\arguments{
\item{silent}{Should rfPIPeak.setup suppress returning the status of installation? Defaults to TRUE.}
}
\description{
This function attempt to check and install the necessary Perl modules and configure data.table for OpenMP
}
\keyword{rfPIPeak}
