% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_annot.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle a vector or matrix for randomization}
\usage{
shuffle(x, along.row = FALSE)
}
\arguments{
\item{x}{input; required.}
\item{along.row}{should the rows be shuffled instead? Defaults to FALSE.}
}
\description{
Randomly rearranges elements in a vector or matrix
}
\keyword{vectors}