% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopbed.R
\name{slopbed}
\alias{slopbed}
\title{Randomly increase/decrease feature lengths in a BED file}
\usage{
slopbed(bed, reformat = TRUE, genome = NULL, read.length = 128,
  slop.range = 300:2000, bin.size = 9, output = NULL)
}
\arguments{
\item{bed}{BED file for input; The file must be tab-delimited and the first 3 columns must conform to BED standards. Alternatively, a data.frame is also acceptable. Required.}

\item{reformat}{should the output be reformatted to a BED6 file including only the newly shifted coordinate information? If FALSE, the resultant feature changes will be appended to the right of the existing columns. Defaults to TRUE.}

\item{genome}{genomic (chromosomal) information corresponding to the BED file. If left as NULL (default), will call fetch.genome() to retrieve precoded hg19 chromosome information, which may or may not be applicable to the supplied BED file.}

\item{read.length}{average read length; this is used to determine the extent of deviations from a feature center, assuming that the length is normally distributed. Defaults to 128 bp, the average read length from our published KR12 Chem-seq results. To turn off random center shifts, set this to 0.}

\item{slop.range}{a vector containing possible span of allowable feature shifts (defaults to 300:2000, based on published KR12 Chem-seq methodology);}

\item{bin.size}{number of randomly selected lengths to change the features (default: 9).}

\item{output}{to write the result to a new BED file, specify the filename here. The default option is not to create a new file and simply return a data.frame to the console.}
}
\description{
Change feature lengths in a BED file based on normally distributed random sizes; the center of a feature is also shifted either left or right at the same time. Inspired by BEDtool's "flankBed" function, this allows one to randomly permute features for various computations.
}
\examples{
set.seed(123)
hg19 <- fetch.genome()
chr <- sample(hg19[,1], 5, replace = TRUE)
x <- round(hg19[match(chr, hg19[,1]),2] * runif(5))
y <- x + 500
slopbed(data.frame(chr,x,y))
## features will be slopped with 9 randomly selected extensions (186 ~ 964bp)
## from both ends, shifted left/right following normal dist. w/ mean = 128 bp
#    chr         x         y           V4          V5 V6
#1  chr8   6667455   6668811 R00051/00428  0.11915888  .
#2 chr20  33283709  33285065 R00014/00428  0.03271028  .
#3 chr11 120481957 120483173 L00071/00358 -0.19832402  .
#4  chrM      8454     10780 R00229/00913  0.25082147  .
#5  chrX  70898462  70899818 R00064/00428  0.14953271  .
}
\keyword{bed}
