% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_list.R
\name{write.list}
\alias{write.list}
\title{Write to a List}
\usage{
write.list(x, filename = "data", append = FALSE, closefile = TRUE,
  outfile)
}
\arguments{
\item{x}{name of the list. Required.}

\item{filename}{name of the list. Required.}
}
\description{
This function write a list to file in plain-text format
}
\examples{
write.list()
}
\keyword{cats}
